refactor(pfb): extract style presets into a usePresets composable
