+++ frappe/public/js/print_format_builder/stores/index.js
@@ -1,5 +1,6 @@
 import { create_default_layout, serialize_layout } from "../utils";
 import { useLayoutHistory } from "./useLayoutHistory";
+import { usePresets } from "../composables/usePresets";
 import { watch, ref, inject, computed, nextTick } from "vue";
 
 // Copy/paste clipboard — persisted to localStorage so a field or section copied
@@ -38,29 +39,6 @@ if (typeof window !== "undefined") {
 
 function clone_plain(obj) {
 	return JSON.parse(JSON.stringify(obj));
-}
-
-const STYLE_PRESETS_KEY = "pfb_style_presets";
-const STYLE_PRESET_KEYS = [
-	"font",
-	"font_size",
-	"label_color",
-	"value_color",
-	"margin_top",
-	"margin_right",
-	"margin_bottom",
-	"margin_left",
-	"page_number",
-];
-function load_style_presets() {
-	try {
-		return JSON.parse(localStorage.getItem(STYLE_PRESETS_KEY)) || [];
-	} catch {
-		return [];
-	}
-}
-function persist_style_presets(list) {
-	persist_local(STYLE_PRESETS_KEY, list);
 }
 
 const SECTION_SNIPPETS_KEY = "pfb_section_snippets";
@@ -475,27 +453,8 @@ export function getStore(print_format_name) {
 		}
 	}
 
-	const style_presets = ref(load_style_presets());
-	function save_style_preset(name) {
-		name = (name || "").trim();
-		if (!name || !print_format.value) return;
-		const style = {};
-		for (const key of STYLE_PRESET_KEYS) style[key] = print_format.value[key];
-		const list = style_presets.value.filter((p) => p.name !== name);
-		list.push({ name, style });
-		list.sort((a, b) => a.name.localeCompare(b.name));
-		style_presets.value = list;
-		persist_style_presets(list);
-	}
-	function apply_style_preset(name) {
-		const preset = style_presets.value.find((p) => p.name === name);
-		if (!preset || !print_format.value) return;
-		Object.assign(print_format.value, preset.style);
-	}
-	function delete_style_preset(name) {
-		style_presets.value = style_presets.value.filter((p) => p.name !== name);
-		persist_style_presets(style_presets.value);
-	}
+	const { style_presets, save_style_preset, apply_style_preset, delete_style_preset } =
+		usePresets(print_format);
 	function find_field_column(df) {
 		const lv = layout.value;
 		const zones = [lv?.header, lv?.footer, ...(lv?.sections || [])].filter(Boolean);
